requirements.txt
#
# note that this file contains requirements for automatic deployments
# Broader requirements are specified in the environment.yml

geographiclib
h5py <=3.12.1
matplotlib <=3.8.4
mpi4py<=4.0.1
openmpi <=5.0.6
numpy<=1.26.4
obspy<=1.4.1, >=1.3.1
pip
prov
pyyaml
scipy <=1.15.0
flake8
tqdm

# Other requirements for tests and docs
flake8
pydata-sphinx-theme
sphinx < 6.0
sphinx-copybutton
sphinxcontrib-mermaid
pytest
py